builtins: add create_tenant(<tenant-name>) overload
This change adds another overload to the `crdb_internal.create_tenant`
builtin that only accepts the name the of the tenant to be created.
The underlying logic automatically picks the next unique ID based on the
existing tenants in the `system.tenants` table. The builtin returns
the ID assigned to the newly created tenant or an error if a tenant
with the same name already exists. Just like other overloads this
buitlin is admin-only and can only be run by the system tenant.

Informs: #91235

Release note: None

+++ pkg/cmd/roachtest/tests/multitenant.go
@@ -29,7 +29,7 @@ func runAcceptanceMultitenant(ctx context.Context, t test.Test, c cluster.Cluste
 
 	const tenantID = 123
 	{
-		_, err := c.Conn(ctx, t.L(), 1).Exec(`SELECT crdb_internal.create_tenant($1)`, tenantID)
+		_, err := c.Conn(ctx, t.L(), 1).Exec(`SELECT crdb_internal.create_tenant($1::INT)`, tenantID)
 		require.NoError(t, err)
 	}
